postcodeの型を統一
